requirements.txt
Django==2.2.4
django-templated-mail==1.1.1
djangorestframework==3.10.2
djoser==2.0.1
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
